requirements.txt
appdirs==1.4.3
click==6.7
Faker==0.7.9
Flask==0.12
Flask-Cache==0.13.1
Flask-SQLAlchemy==0.16
futures==3.0.5
gunicorn==19.7.0
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
redis==2.10.5
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.12
